Extract DimensionMode into its own file
